requirements.txt
pyautogen==0.7.1
fastapi==0.116.1
uvicorn==0.34.0
python-dotenv==1.0.1
openai==1.109.1
PyPDF2==3.0.1
python-multipart==0.0.19
faiss-cpu==1.13.2
numpy>=1.24.0,<2.0.0
nanoid==2.0.0
autogen-agentchat==0.7.5
autogen-ext==0.7.5
gunicorn